Hacer boletin3

+++ Tema2/src/boletin1/Ejercicio8.java
@@ -6,15 +6,15 @@ public class Ejercicio8 {
 
 	public static void main(String[] args) {
 		
-		// Creamos una variable para almacenar la nota que introducirá el alumno teniendo en cuenta que la nota es un número entero
-		int nota;
+		// Creamos una variable para almacenar la nota que introducirá el alumno teniendo en cuenta que la nota puede tener decimales
+		double nota;
 		
 		// Creamos el scanner
 		Scanner reader = new Scanner(System.in);
 		
 		// Pedimos al usuario que introduzca la nota mediante un syso
 		System.out.println("Introduzca su nota");
-		nota = reader.nextInt();
+		nota = reader.nextDouble();
 		
 		// Si la nota está comprendida entre el 0 y el 10 que compruebe la calificación del alumno, sino que muestre un aviso al usuario por consola
 		if (nota >= 0 && nota <= 10) {
